[AGP] Register generated kotlin sources with sources.java API
